+++ geoimagenet_ml/store/datatypes.py
@@ -21,7 +21,7 @@ from geoimagenet_ml.utils import (
 from geoimagenet_ml.status import COMPLIANT, CATEGORY, STATUS, job_status_categories, map_status
 from geoimagenet_ml.processes.types import process_mapping, PROCESS_WPS
 from geoimagenet_ml.store import exceptions as ex
-from geoimagenet_ml.ml.impl import load_model, valid_model
+from geoimagenet_ml.ml.impl import load_model, validate_model
 from pyramid_celery import celery_app as app
 # noinspection PyPackageRequirements
 from dateutil.parser import parse
@@ -717,7 +717,7 @@ class Model(Base, WithName, WithUser, WithVisibility):
         if not success:
             msg = "Failed loading model data: [{!r}].".format(exception or "unknown reason")
             raise ex.ModelLoadingError(msg)
-        success, exception = valid_model(data)  # nothrow operation
+        success, exception = validate_model(data)  # nothrow operation
         if not success:
             msg = "Failed validation of model configuration: [{!r}].".format(exception or "unknown reason")
             raise ex.ModelValidationError(msg)